\name{demo01-package}
\alias{demo01-package}
\alias{demo01}
\docType{package}
\title{\packageTitle{demo01}}
\description{\packageDescription{demo01}}
\section{Package Content}{\packageIndices{demo01}}
\author{\packageAuthor{demo01}}
\section{Maintainer}{\packageMaintainer{demo01}}
\keyword{package}
