% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{The Hello Function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
The Hello Function
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}
}
